requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pydantic>=2.8,<3.0

langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-community>=0.3.0,<0.4
langchain-groq>=0.2.0,<0.3

langchain-chroma>=0.1.0
chromadb>=0.5.5
langchain-text-splitters>=0.0.2

sentence-transformers==3.0.1
pdfminer.six==20240706
python-docx==1.1.2
tqdm==4.66.4
langchain-huggingface>=0.0.3
langchain-groq>=0.2.0,<0.3
